+++ derive/src/subscription.rs
@@ -152,6 +152,7 @@ pub fn generate(
                     default_with,
                     validator,
                     visible: arg_visible,
+                    secret,
                 },
             ) in &args
             {
@@ -194,6 +195,7 @@ pub fn generate(
                         default_value: #schema_default,
                         validator: #validator,
                         visible: #visible,
+                        is_secret: #secret,
                     });
                 });